quit the loop if the indexer were aborted

+++ webcurator-core/src/main/java/org/webcurator/core/store/PywbIndexer.java
@@ -165,6 +165,9 @@ public class PywbIndexer extends IndexerBase {
                 return;
             }
             for (File warc : warcFiles) {
+                if (!this.isRunning) {
+                    break;
+                }
                 File archivedWarc = new File(pywbManagerStoreDir, "archive" + File.separator + warc.getName());
                 if (archivedWarc.exists()) {
                     boolean ret = archivedWarc.delete();
